Sprint04

+++ Encargado/eliminarCategoria.php
@@ -1,13 +1,18 @@
 <?php
     include("seguridad.php");
     include("../conexion.php");
-    $cod = $_GET['cod'];
+    if($_SERVER['REQUEST_METHOD']==='GET'){
 
-    $query = "DELETE FROM categoria WHERE idCategoria=$cod";
-
-    mysqli_query($conn,$query);
-
-    mysqli_close($conn);
-
-    header("LOCATION:categorias.php");
+        $cod = $_GET['cod'];
+        
+        $query = "DELETE FROM categoria WHERE idCategoria=$cod";
+        
+        mysqli_query($conn,$query);
+        
+        mysqli_close($conn);
+        
+        header("LOCATION:categorias.php");
+    }else{
+        header("LOCATION:../cerrar_sesion.php");
+    }   
 ?>
